direct, istream-pipe: re-enable SPLICE_F_NONBLOCK
splice() without SPLICE_F_NONBLOCK may cause istream_pipe to hang,
when you pass a length to splice() which is bigger than the pipe
buffer.  So just assuming that splice() won't block because is pipe
buffer is empty is not enough.

Now that istream_file is safe against the Linux kernel NFS+EAGAIN
problem, we can reenable SPLICE_F_NONBLOCK.