fork: re-add event after some input data has been read
This fixes a bug where xsltproc's stdin is starving because beng-proxy
discontinues to send data.  This is because beng-proxy has removed its
"write" event on the stdin pipe, and expects the connected istream to
"call back" when new data arrives.  This however never happens when
istream_file is used, because istream_file doesn't use any events, and
thus cannot call back.