http-server: don't set up event in http_server_consume_body()
At the end of this function, the buffer is never full (if it were,
http_body_consume_body() would have returned 0, and we would have
returned already).  All callers will set up the event after a
successful http_body_consume_body() call, so we don't need to do it
here.