http-server: check if client closes connection while processing
Continue reading the next pipelined request into the buffer after the
current one has been finished.  We are now able to detect when the
client closes the connection, and we can cancel the request.  If the
pipelined request is too large for the input buffer, keep-alive is
simply disabled and the pipeline is discarded.