istream-replace: update the buffer reader after new data was added
When data is appended to the growing_buffer after the reader has been
moved to the (old) end-of-buffer, the reader structure becomes
inconsistent and violates its own rules, causing an assertion
failure.  This patch adds a new method called _update(), which checks
for this condition, and should be called after each append.