file-handler: handle If-Modified-Since followed by filter
A filter is supposed to be idempotent, and if a file was not modified,
the filtered result must not be modified either.  So it's legal and
useful to handle If-Modified-Since in the file handler, even if it is
followed by a filter.  Note that the processor is not a filter, and is
not affected by this patch.