pool: use new dedicated area for large allocation
Further reduction of overhead: if a large allocation is requested that
does not fit in the current area, a new one is allocated, and
remaining memory in the current area is never used.  This patch keeps
on using the new area, but allocates a dedicated large area for the
requested allocation.