cgi: fix spontaneous shutdown due to misrouted SIGTERM signal
On a busy production server, the worker process was often shut down by
SIGTERM, but it looked like nobody sent it.  At the same time, a CGI
process hung, blocking the worker shutdown forever.

The problem was a race condition due to how libevent implements signal
handlers: a child process was forked, and right after that, beng-proxy
changes its mind and decides to kill the new process.  Problem is that
this went too fast, and the new process did not yet get around to
execute the CGI program, and libevent's signal handler is still in
effect, which sends the number 15 (SIGTERM) to the socket pair, to be
received by the original beng-proxy worker process, which then shuts
down.

This race condition can only be solved by suppressing signals while
the transition happens, and cleaning up the signal handlers in the
child process.