file-handler: fix xattr usage on ranged file request
For a ranged request, istream_file was wrapped inside istream_head,
and the latter was casted to istream_file for the istream_file_fd()
call.  This caused EBADF errors and possibly assertion failures, in
case the malformed file descriptor happened to be negative.

This fix adds a new method called _set_range(), which invokes lseek(),
replacing istream_skip(), and limits the number of bytes istream_file
will deliver.  This way, only the data that will be needed will be
read.  This allows dropping istream_head.