namespace_options: work around uid/gid mapper failure using PR_SET_DUMPABLE
Problem was that open("/proc/self/uid_map",O_WRONLY) failed with
EACCES, because the file was still owned by root.  The proc filesystem
assigns the task's uid to /proc/self/* only if the task is "dumpable",
and the "dumpable" flag gets reset by setreuid().  We use
PR_SET_DUMPABLE to restore the flag during startup.