buffered_socket: ScheduleRead() defers call to Read()
When the caller indicates that it wishes to read from the socket,
schedule a call to Read() instead of waiting for more data on the
physical socket.  If all data has been received into the input buffer
already, more data will not appear on the socket, thus starving the
input stream.  Instead, we should deliver what we already have.

This fixes problems with large SSL packets: all data has been received
already into the input buffer, most of it (but not all) has been
copied to the SSL filter's buffer, and now it's waiting for more -
which will never be delivered, because no more data is received from
the buffer, and BufferedSocket's input buffer will not be considered
again.