was_client: hold caller_pool reference while sending response
During http_response_handler::response(), the caller must hold a
reference to the pool that was passed to was_client_request().  The
reference to was_client's pool was there to be able to check if the
was_client was destroyed after returning, but a reference to the other
pool was missing.

Bug was revealed by commit 856ca73947c3033886c560af839aba9d60509e5c