ThreadSocketFilter: no CycleBuffers() in Done()
This call has always been useless, because it does not solve the
problem with unmovable allocations in a nearly-empty SliceArea.

The real short-term solution is to cycle the buffers on demand,
i.e. in the fb_pool cleanup timer.  The long-term solution (for v11)
will be to always free all empty I/O buffers.