request_session: delete the "session" parameter if client is bot
If a bot sends a request with a "session" parameter (valid or not), it
gets ignored, but yet propagated to processed URIs, because it remains
in the "args" map.  This is bad, because it multiplies the number of
bot requests for each session id a bot happens to stumble upon.