thread_socket_filter: cancel the handshake_timeout_event on disconnect
If the handshake_timeout_event fires while the worker thread is
currently waiting for the certificate database, we get a crash.  So as
soon as the client disconnects, we should disable the
handshake_timeout_event, because without a client, the handshake
cannot ever be finished, and all we're waiting for is internal.