ThreadSocketFilter: set the "want_write" flag on short write
If all of the data could not be written to the buffer, then the caller
will not feel responsible for calling again; we need to remind him
we're ready for writing later.  The way to do it is to set the
"want_write" flag which will attempt another write later, when the job
has finished.