lb/ForwardHttpRequest: use FNV-1a hash instead of DJB hash
The DJB "cdb" hash performs very poorly when only the last byte of the
input is changed, which however is a common use case for sticky=host.
The best solution would be to switch to a cryptographic hash, but that
is expensive at runtime, and not worth the trouble.  It turns out that
FNV-1a performs well enough for this use case, while being just as
fast as DJB hash.