shrink I/O buffers back to 8 kB
Reverts commit d2bfa3ebadf because it appears to be no longer
necessary, and maybe it was never really necessary: OpenSSL allocates
an internal buffer large enough for the maximum record size, and our
FifoBuffer doesn't need to be able to hold all of that.

Maybe the larger buffer size was a kludge to work around the bugs
which were fixed by 41b02071685 and c8e0fe7cd10?