ThreadSocketFilter: fix receive stall bug
When checking for a stalled receive filter, it does not matter if
there is data in the encrypted_input buffer; it only matters if the
decrypted_input buffer is full.

Previously (when I wrote commit d2bfa3ebadf), OpenSSL's behavior was
different and this check was correct, but apparently newer OepnSSL
versions have changed slightly.