net/PConnectSocket: hold pool reference while invoking handler
If the connect attempt finishes immediately, we must invoke the
handler.  The handler however assumes that we're holding his pool
reference.

For example, the SendError() call in LbRequest::OnStockItemError()
could finish immediately as well, destroying the last reference to the
request pool, crashing ResponseSent() due to use-after-free.

And wow, this bug is 10 years old: 6bfe62f2cfbd73b6