fs/Lease: move buffers and destroy socket in Release()
This is needed for persistent HTTPS client connections: Release() must
allow others to reuse the FilteredSocket instance, and thus we have to
adopt its buffered data and dispose it.

This is just a test ride for the real implementation; later, we will
move to FilteredSocket back to its stock instead of destroying it.