istream/subst: optimize mismatch check
Previously, each occurrence of a start character caused SubstIstream
to flush all data, changing to State::MATCH, only to roll back if the
following character (which may have already been known) mismatches.
This comes with a lot of overhead: too many HTTP chunks are generated,
and thus too many system calls are invoked.

This commit introduces the CheckMatch() function which checks if the
remaining data in the input buffer really matches before changing to
State::MATCH.