http_request: abolish the forced "Connection:keep-alive" request header
This header was never necessary to enable keep-alive, because that was
the default in HTTP/1.1.  In 2007, when I wrote this code line (commit
343254dd2c), I thought it would improve compatibility with HTTP/1.0
servers.  But who really cares about HTTP/1.0 servers...

The problem is that WebSocket's use of the "Connection" header could
lead to a duplicate "Connection" header, which is illegal.  This could
break the WebSocket handshake.