bp/ForwardHeaders: fix WebSocket header duplication
WebSocket headers were sent twice if the header forward setting
"OTHER=yes" was given, because WebSocket headers are considered to be
in group "OTHER".  This commit adds additional checks to avoid
forwarding those headers again if "OTHER=yes".  A similar change had
to be made for group "CORS", which includes the request header
"Origin", also part of the WebSocket specification.