shm/dpool: add offset to size check, fixes assertion failure
On a server, crashes were observed with allocation sizes of 3914
bytes.  That is smaller than `managed_external_buffer::get_size()`,
but bigger than the largest allocation which fits into the
`managed_external_buffer`.  Experiments showed that the offset between
`get_size()` and the largest possible allocation is 152 bytes, but
there appears to be no way to query this.  However,
`segment_manager::get_min_size()` (192 bytes) seems to be large enough
to act as an offset, so let's use this one to fix the assertion
failure at the end of `dpool::Allocate()`.