fb_pool: enlarge I/O buffers to 32 for faster bulk transfers
Originally, the size was 8 kB and was enlarged to 16 kB in commit
d2bfa3ebadf700b4af486614a48e05fab6941cc5 to work around an OpenSSL
problem.  When that one was fixed, the size was shrunk back to 8 kB in
commit cb941ced5d2cc7d32b5fb59f852f7d1d50cd5d0a.

For large transfers, it may be helpful to copy larger chunks at a time
to reduce CPU overhead.  Since beng-proxy has become pretty good at
releasing empty buffers quickly, the memory overhead for larger
buffers is very low.  So let's experiment with 32 kB buffers.