widget/Request: widget tag headers replace existing headers
This was requested by widget developers so they could pass trusted
information from the containing template to the widget.

Strictly speaking, this is an API change, but it was assumed that no
widget ever relied on sending both client's and templates' headers.
On the other hand, this feature change gives a stronger guarantee
about what happens.