istream/pause: automatically call Read() again after Resume()
Without this, some Istreams stall forever because there never was a
Read() call for them; the only one was blocked by
PauseIstream::_Read().

This could happen with focused inline widgets when the request body
was tiny enough to fit in the first slice buffer.  This way, there was
never another `EPOLLIN`.