debian/libcm4all-beng-proxy3-dev.install: copy from debian/tmp/
In commit a203c0a1bf6, the include directory in the pkg-config file
was changed to "-3", but the headers were still installed to "-0".
This was because the path was hard-coded in
debian/libcm4all-beng-proxy3-dev.install as well, but this file wasn't
changed.

To make this less fragile, this commit eliminates this explicit
reference and uses "ninja install" instead to feed everything from one
source (the `--includedir=...` option in `debian/rules`).