was/Client: send PREMATURE in Cancel() and WasInputClose()
If our client has canceled the request, we need to tell the WAS
process how much request body data has been sent already and that it
shall stop expecting more data.  This prevents a bug that would kill
the WAS process because it kept on reading from the request body, not
knowing that this request body had been canceled.