net/FailureInfo: add counter for protocol failures, disable after several failures
If there is a protocol-level error on a HTTP connection (e.g. a
connection reset), don't disable the node just yet (because that will
invalidate all sessions on that node); instead, wait for 8 or more
consecutive failures before disabling the node.  Each successful
request will reset the counter.  This way, sporadic failures will not
lead to session loss.