bp/CssProcessor: swap tee.first and tee.second
Like commit 57ee0c54d7de7e eight years ago, this one sends data to the
ReplaceIstream first to fill its buffer, and then to the CSS parser.

This fixes another use-after-free bug of the WidgetContext detected by
PoolLeakDetector, because CssProcessor holds a reference to the
WidgetContext which must be released before the parent pool is
destroyed.