http_request: move the StringMap
Another regression from commit
33cc95991fca5017382bfad9361665bd8c1330a1: by keeping a reference to
the `StringMap`, we run into a use-after-free bug, because some
callers (such as `WidgetRequest::SendRequest()`) pass a temporary
`StringMap` to `ResourceLoader::SendRequest()` which goes out of scope
quickly, but the reference is used later, when the connection to the
HTTP server is established.

So instead of holding a reference, move the StringMap to a local copy.