bp/Instance: remove support for multiple worker processes
After 15 years into this project, we have never used the feature.
beng-proxy has always been a single-threaded, single-process web
server, and the world around it has been the limiter.

If we ever feel we need that back, we might consider moving some
connections to separate threads.  But for now, carrying around this
multi-process feature looks like a burden.