istream/buffered: read from file descriptor to buffer
An assertion failure has occurred on several servers where the
http_client had delivered the first few bytes in a buffer, and then
tried to submit a file descriptor for splicing.  Upon seeing a buffer,
BufferedIstream::OnData() attempted to disable the splice by calling
`input.SetDirect(0)`, but the change was not propagated to all
callers, leading to an `BufferedIstream::OnDirect()`, which crashed
due to `assert(!buffer.IsDefined())`.

The solution is to just use `read()` to read from the given file
descriptor if a buffer was already allocated.