bp/errdoc: refactor to Coroutine
The first coroutine in this project.  This is a method which was easy
to refactor, but the API is a mess, because it picks up the original
request in the `co_response` field, and needs to postpone the
DispatchResponse() call to the Co::InvokeTask completion function,
outside of coroutines.  This will be cleaned up when there is a more
consistent concept of how coroutines will be integrated in request
handlers.