nghttp2/{Client,Server}: remove END_STREAM check from OnDataChunkReceivedCallback()
In this libnghttp2 callback, the frame payload may not have been
received completely, so after pushing the partial data to the buffer,
we must not yet signal end-of-stream, because more data is expected.

We may check for the END_STREAM frame flag only after the frame
payload has been fully processed.  And this code already exists in
OnFrameRecvCallback(), leading to end-of-stream notification in
Request::OnEndDataFrame().

The simple solution is to just remove the redundant (and misplaced)
code from OnDataChunkReceivedCallback().  This eliminates the need for
the "eof" field, because there is now just one entity checking for
end-of-stream.