stock/MultiStock: add lease immediately in OnStockItemReady()
Fixes a race condition: if Get() is called before the DeferEvent
fires, the new Item goes to the Get() caller, not to the Waiting
instance, whose Request object has already been consumed.  Once the
DeferEvent fires, the assert(w.request) fails, and we have no chance
to submit a new request.