stock/MultiStock: manage an inner layer of StockItem instances, no Lease
This finally fixes the concurrency limit problems with the old
implementation, which was caused by the MapStockMap sitting in front of
the MultiStock, whose Stock instances could not be integrated into
MultiStock's "waiting" list.

This commit eliminates this StockMap instance and integrates its
features.  As a side effect, this makes users of MultiStock much
easier to use, because lots of previously redundant code now lives
inside class MultiStock.