ssl/CompletionHandler: add flag to fix race
If a certdb callback gets queued in the worker thread right after the
main thread cancels the SslFilter, the certdb lookup continues as
usual, because the worker thread does not know the SslFilter has been
cancelled already.  This commit adds a flag which remembers this
state protected by the newly introduced mutex.