nghttp2/Server: fix stall bug with many concurrent uploads
When there were many concurrent uploads, but some of them were stalled
(e.g. because the WAS concurrency limit was reached or because we're
waiting for "100 Continue"), all pending uploads congested not only
the stream-level window (which is the whole point of the window), but
also the connection-level window, stalling all request body uploads of
all streams.  Since all "active" streams were also stalled, there was
no chance for progress on those, and no WAS concurrency slots would
ever be freed, effectively deadlocking the whole HTTP/2 client.

The solution is to start with a very low stream-level window size, and
increase this window size only after we have figured out that this
request body is going to be used.  This will still add a bit of
connection-level congestion, but not enough to fill it completely.