nghttp2/Server: always update the connection-level window
Always call nghttp2_session_consume_connection() for data that was
received, to keep the connection-level window open for more data all
the time, to avoid all transfers getting stuck because the
connection-level window was congested with request bodies that are not
used currently.

Switch Request::Consume() to nghttp2_session_consume_stream(), which
updates just the stream-level window.  This means we can omit all
Consume() calls after the end of a stream - why bother update the
stream-level window if we're not going to use it anyway?