fs/Factory: split class SocketFilterParams
Being independent of the memory pool fixes a use-after-free bug in
class Nghttp2::Stock: if the GetRequest that created the Item got
canceled, then its SocketFilterFactory got freed already, but after
ConnectFilteredSocket() finishes the TCP connect, it attempted to use
that SocketFilterFactory.  This commit solves the problem by using the
light-weight class SocketFilterParams until we figure out that we need
a new connection, and then create the SocketFilterFactory, moving all
data from the pool to the normal C++ heap.  This way, the initial
GetRequest can be canceled but the SocketFilterFactory remains valid,
because it is now owned by class ConnectFilteredSocketOperation.