test/t_client: disable several bucket-related checks for FastCGI
The FastCGI protocol doesn't have a "Content-Length" header which
allows announcing the total length in advance, and therefore the
GetAvailable() implementation works only if the "END" packet has been
received already; but this unit test was very flaky because sometimes,
it received only the first packet while the rest was not yet sent by
the dummy FastCGI server.  This is difficult to get right, so let's
just disable those tests on FastCGI for now.