istream/Subst: enable fallback mode in one last code path
SubstIstream::_FillBucketList() is not yet implemented completely, but
one code path was missing for enabling the fallback mode, thus making
the caller believe that it just needs to call _FillBucketList() again
to receive more data, which however did not work and could lead to
stalled transfers.  This bug usually didn't take effect, until a
cached widget was requested in "frame" mode.