http_client: disable keep-alive if writing request body fails early
If writing the request body fails early, then `keep_alive` will be set
to `false`, but after that, it will be set to `true` again in
HeadersFinished().

Solve this by initializing the field with `true` and only set it to
`false` when `Connection:close` is found, but never revert it to
`true`.

This fixes a spurious unit test failure.