lb/Cluster: improve Rendezvous Hashing, combine hashes without XOR
Combining the two hashes with XOR resulted in a strong bias towards
some servers, proportional to the distance between the server hashes
(and nothing else).  Turns out the short cut using XOR doesn't work,
and we need to really calculate the hash of server address and
request.

The new implementation keeps the precalculated address hash and then
uses it as initial value for the DJB hash calculation of the request's
"host" header (or whatever stickiness is configured).  This properly
scrambles the hash and results in an equal distribution independent of
the server address hash distances.

Interestingly, doing it the other way round revealed a weakness of the
DJB hash algorithn: instead of 2.6% difference between highest and
lowest member, the difference rocketed to 63%.  This problem was not
present in libsodium's generichash (BLAKE2b): there, I could observe
2.8% and 1.4%.  FNV revealed the same weakness as DJB: 2.3% and 18%.

Since 2.6% is good enough and DJB is the fastest hash function, I
decided to simply do DJB_HASH(address+host).  This however required an
API change: previously, the host header was passed as a hash
(parameter `sticky_hash`); now we need to pass the actual host header
as well (new parameter `sticky_source`).