http/local/Glue: ReleaseLease() always returns PutAction::DESTROY
The `LhttpLease` class owns the `FilteredSocket` (and
`BufferedSocket`) instance, and it is destroyed in ReleaseLease().
This makes BufferedSocket::InvokeData() expect that the handler
returns BufferedRsult::DESTROYED, but since PutAction::REUSE is
returned, `FilteredLease` does not know that the `FilteredSocket` was
destroyed.

This commit is a kludge to work around the assertion failure.