io/UringOpenStat: use RESOLVE_BENEATH of dfd==-1, not AT_FDCWD
A side effect of commit dc8941f1879388 was the change of
`handler.file.base` from `AT_FDCWD` to -1 if there was no BASE.  Using
-1 is more correct, however that change broke the `RESOLVE_BENEATH`
check in the `UringOpenStatOperation` constructor, which did not make
sense anyway.  To fix this, we now check `FileDescriptor::IsDefined()`
and enable `RESOLVE_BENEATH` even for `AT_FDCWD` (which we never use,
anyway).