fs/Stock: keep (and reuse) canceled outgoing connections
If an incoming HTTP request triggers a connect to the backend server
and it gets canceled before the connection is established, we would
close the connection, even if somebody else is also waiting for the
same connection.  This is an unnecessary waste of resources.  Instead,
let's keep connecting and when it finally finishes, add it to the
stock via InjectIdle(), to be used by whoever wants the connection
next.