bp/Base: do not open the BASE directory if BENEATH is used
If a BENEATH directory has already been opened, don't open the BASE
directory as well.  The protection from RESOLVE_BENEATH on the BENEATH
directory is enough, but keeping the BASE directory in the FdCache as
well might cause trouble because the user could rename/delete that
directory, leaving the FdCache with stale entries for longer than the
user is willing to wait.

This eliminates one file descriptor, but requires StripBase() to copy
the path.

As a side effect, (relative) symlinks across DocumentRoots are now
possible.