http/cache: store "no-cache" responses (and replace existing item)
A request with "Cache-Control:no-cache" is supposed to revalidate all
cache items, but the response may be stored.  In fact, the response
should be stored, because this new response should really invalidate
a response that was cached previously.  Therefore, do not skip the
cache completely on "no-store", but instead force revalidation with
the origin server.