ssl/Filter: clear the `drained` flag after moving data to buffer
The ThreadSocketFilter flag `drained` specifies whether the handler's
internal buffers are empty.  As long as there is still data in the
internal buffers, the FilteredSocket cannot be shut down at the end of
a non-keep-alive HTTP response.  The call to
FilteredSocket::InternalDrained() (which then shuts down the TCP
socket) will be deferred until the flag becomes true and all other
buffers (the ones in ThreadSocketFilter) are flushed to the TCP
socket.

There are two callers of InternalDrained(): one in
ThreadSocketFilter::Done() which gets called after the SslFilter
finishes in the worker thread, and another one in InternalWrite()
which flushes buffers to the TCP socket.

There is only one place where this `drained` flag gets set: at the end
of SslFilter::Run() when the internal buffers may have been moved to
the external buffers (`ThreadSocketFilter::encrypted_output`).
However, at the beginning of Run(), the internal `plain_output` buffer
get filled (moved from ThreadSocketFilter's `plain_output` buffer).
There, the `drained` flag was not cleared.

Depending on the timing, a call to ThreadSocketFilter::InternalWrite()
may then find `drained==true` even though the SslFilter is currently
running in the worker thread with non-empty internal buffers.  If
there also happens to be a pending socket shutdown, InternalWrite()
may then execute this shutdown before the rest of the response (that
is currently being encrypted) was sent.

The fix is trivial: set the `drained` flag at the begining of
SslFilter::Run() as well.

This bug is now ten years old: it has existed ever since the `drained`
flag was implemented by commit 1e375e3a8c6a5782321be4b5 (2014/03) -
but it was never noticed because (almost) nobody disables HTTP/1.1
keep-alive.

To sum it up, the bug occurred when all of the following conditions
were met:

- HTTP/1.1
- keep-alive disabled
- TLS encryption
- very fast HTTP response handler (e.g. from cache)
- big HTTP response (> 64 kB)
- socket output buffers empty, internal TLS buffer non-empty
- socket becomes writable exactly while encryption runs in worker thread