bp: continuously watch memory usage after memory warning
Implement a margin; until the memory usage falls below that margin,
processes get stopped repeatedly (even if we're already below
MemoryHigh) to avoid further kernel shrinker calls due to MemoryHigh.
The problem is that the shrinker is unable to free memory; it's mostly
anoymous memory (some of it shared).  The kernel can't do much here.