{bp,lb}/ListenerConfig: increase backlog to 4096
This massive increase from 64 to 4096 is useful to avoid accidental
syncookies which reduces network performance because syncookies means
we lose information about TCP options requested by the client.  A very
insightful blogpost explaining this can be found here:

 https://blog.cloudflare.com/syn-packet-handling-in-the-wild

The old value of 64 was ridiculously small, and such a small SYN queue
would overflow too often.  4096 is Linux's default value for
/proc/sys/net/core/somaxconn and thus seems reasonable for servers
such as beng-proxy and beng-lb.