ResourceAddress: do not copy objects in Expand()
Expand() is supposed to modify the `*this` object.  Why do we copy the
HttpAddress/CgiAddress/... objects first?  This used to be necessary
long ago, but hasn't been for years.

This method is only called from the translation cache, right after
copying the ResourceAddress with CacheLoad().  We are already
operating on an isolated ResourceAddress instance sharing no pointers
with any other instance, so this is safe.

This removes a lot of CPU and memory overhead from the translation
cache.