istream: remove method AsFd()
I have been carrying this around for a long time, but it was a
CGI-only optimization for a rare corner case, but it's not worth
carrying the code cruft around for this.  It doesn't work well when
only a portion of a file is to be served and doesn't work at all with
io_uring.  Let's remove it!