io/FdCache: allow trailing slash in "strip_path" parameter
When passing the BASE path to FdCache::Get(), StripLength() always
returned 0 because the BASE ends with a slash, but StripLength()
expects the "path" parameter to have another slash.  This caused
stripping to fail, leading to EXDEV errors due to RESOLVE_BENEATH.

This bug occurred since commit
3570067b9dfb825c374aa545f1512043e0798e10 when BASE was used with
FdCache for the first time.