istream/UringSplice: restart the splice operation after EAGAIN
Fixes another stall bug that could occur if the pipe was nearly full
(less than one page of space).  In that case, the splice operation
fails with EAGAIN and nobody ever cares to restart the operation.  We
should do that if we're confident that the pipe is no longer full -
i.e. if _ConsumeDirect() has been called again in the meantime.