istream/UringSpliceIstream: move the PipeLease into the inner struct
This fixes a rare bug that occurs when a io_uring splice() got
canceled before it was submitted to the kernel; the old file
descriptors were still in the io_uring, but if they happened to be
pointing to another pipe meanwhile, that other pipe got clobbered with
data from the previous file.  That could cause data corruption and
assertion failures.

By moving the PipeLease into the SpliceOperation struct, it will
remain valid until the kernel completes the io_uring operation.