ResourceAddress: fix off-by-one bug in uri_path_verify_paranoid() call
We always passed "tail-1" to uri_path_verify_paranoid() because we
wanted to pass the preceding slash, which was guaranteed to exist; but
after normalization was implemented by commit ffeb7e2694fe6, "tail-1"
could be an out-of-bounds pointer.

Since uri_path_verify_paranoid() works just as well with the non-slash
pointer, we can do that; but we need to add a check for another
leading slash, because uri_path_verify_paranoid() disallows double
slashes, and omitting the first one would break the check.