istream/replace: remove empty substitutions in _FillBucketList()
Fixes a stalled transfer when an empty substitution (istream_null) is
at the front of the list and an incomplete widget follows directly
after that.  The _FillBucketList() call will return no data, thus
_ConsumeBucketList() does not get called and the empty substitution is
not removed.  The next Substitution::OnIstreamReady() call has no
effect because it's not the active substitution.

To fix this, we need to remove empty substitution in _FillBucketList()
without relying on _ConsumeBucketList() to be called eventually.