istream/replace: reset the "more" flag before recursing
Fixes regression from commit e105d8b3bba2f86abe9ad6b00128d7af78072e6f
which caused an assertion failure.  We need to restore the
IstreamBucketList to the state before this method was entered before
doing the recursive call (which was supposed to be a temporary
workaround anyway, but an ugly one).