istream/replace: pop all active substitutions that have no input
This patch adds ToNextSubstitutionLoop(), which calls
ToNextSubstitution() in a loop while the front-most substitution is
active and has no input.  This fixes another stall bug that can occur
if a non-front substitution receives EOF but never becomes active and
never continues the transfer when after that, the front-most
substitution receives EOF as well.  In that case, all empty
substitutions must be popped.