istream/{concat,replace}: handle FALLBACK on input's OnIstreamReady()
If the input that has invoked OnIstreamReady() has been closed
already, it will never see FALLBACK and thus will never handle it.
And even if it did, it would only invoke the fallback code on itself
and not on the Istream that requested the fallback.

Two very old bugs revealed by the optimizations from version 21.0.
Previously, that bug had been invisible because excessive EPOLLOUT
events would eventually continue the transfer, but these EPOLLOUT
registrations have been optimized away in version 21.0.