http/server/SRequest: implement DechunkHandler::OnDechunkEnd()
Update internal book keeping and destroy the HttpBodyReader which is
going to be abandoned by the DechunkIstream.  This fixes an assertion
failure (in HttpServerConnection::CloseRequest() on
"request.read_state != Request::BODY") that can occur when the end
chunk was handled and the incoming HTTP connection gets closed by the
client before the response is finished.