http/server: fix DechunkInputAction::ABANDON handling
This amends the broken bug fix commit
0b92a0084337d72851 by adding state ABANDONED_BODY.  Instead of
destroying the RequestBodyReader from within FeedRequestBody() (which
causes other bugs), this is postponed to CloseRequest(), and the new
state ABANDONED_BODY keeps track of this.

This keeps the RequestBodyReader in a well-defined non-destroyed state
which allows it to handle the return values from DechunkIstream.  It
is destroyed when it is safe to do so, in a different stack frame.