http/Client: add missing empty() check
The preceding InvokeReady() call may have emptied the buffer.  If it
then returns FALLBACK, for example to switch to splice() after parsing
response headers from the buffer, HttpClient::OnBufferedData() may
then call FeedBody() with an empty span.  This causes an assertion
failure because feeding an empty span is not allowed.

The same check already existed in the HTTP server since commit
c19769c71c5f4cb, but I forgot to add it to the HTTP client as well.