http/{client,server}: OnBufferedData() returns MORE if needed
Yet another old stall bug revealed by the 21.0 optimizations: when
OnBufferedData() consumes some data (with TLS, from
ThreadSocketFilter::unprotected_decrypted_input) and the whole rest of
the body is already in ThreadSocketFilter::decrypted_input, nobody
will ever bother to use that because IstreamBucketList's "more" field
was PUSH - a promise that the consumer would be called back as soon as
more data is available.  We can't return PULL or AGAIN instead because
that piece of data may arrive concurrently from the worker thread.

The only proper solution is to return BufferedResult::MORE in such a
case which will cause ThreadSocketFilter::SubmitDecryptedInput() to
call OnBufferedData() again.

Prior to version 21.0, the consumer would always schedule a write
event (EPOLLOUT) which would pull the remaining data from the body.
The point of the 21.0 optimizations was that the epoll_ctl() call
adding EPOLLOUT was not necessary because more data is already
available, but that requires getting the procuder to behave correctly.