fs/ThreadSocketFilter: add `input_serial` to track new data
The method SubmitDecryptedInput() needs to know whether new data was
added to `unprotected_decrypted_input` by MoveDecryptedInput().  This
currently works by checking whether its GetAvailable() return value
increases, but that works only if MoveDecryptedInput() was never
called from inside InvokeData().  To make this more robust, use a
serial that gets incremented for the check.

This fixes a stall bug in the HTTP server with TLS.